+++ src/header.rs
@@ -3,17 +3,17 @@
 use serde::{Serialize, Deserialize};
 use std::time::SystemTime;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum FileType {
     Dir,
     File,
     SystemLink,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Metadata {
     file_type: FileType,
-    size: u64,
+    size: u128,
     permissions: u8,
     modified: Option<SystemTime>,
     accessed: Option<SystemTime>,
@@ -36,11 +36,15 @@ impl From<std::fs::Metadata> for Metadata {
     fn from(data: std::fs::Metadata) -> Self {
         Metadata {
             file_type: data.file_type().into(),
-            size: data.len(),
+            size: data.len() as u128,
             permissions: if data.permissions().readonly() { 1 } else { 0 },
             modified: data.modified().ok(),
             accessed: data.accessed().ok(),
             created: data.created().ok()
         }
     }
+}
+
+impl Metadata {
+    pub fn len(&self) -> u128 { self.size }
 }