+++ src/safepath.rs
@@ -1,0 +1,5 @@
+
+//A wrapper around Path of PathBuf that prevents invalid or unsafe archive paths (like the infamous '../' component)
+struct SafePath {
+
+}